refactor: organized files and created test data source interface.
