+++ src/Beta/ITestDataSource.cs
@@ -1,0 +1,6 @@
+namespace Beta;
+
+public interface ITestDataSource<out T> : IEnumerable<T>
+{
+    bool SupportsEagerLoading { get; }
+}
