+++ src/Beta/RawEnumerableTestDataSource.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+
+namespace Beta;
+
+public class RawEnumerableTestDataSource<T>(IEnumerable<T> scenarios) : ITestDataSource<T>
+{
+    public bool SupportsEagerLoading => true;
+    public IEnumerator<T> GetEnumerator() => scenarios.GetEnumerator();
+    IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+}
